requirements_langchain.txt
# LangChain Agent Dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
sentence-transformers>=2.2.0

# Optional: For advanced agent features
langgraph>=0.1.0  # For complex agent workflows
langsmith>=0.1.0  # For agent monitoring and debugging


